Refactored MicToSpeechly to use EnergyTresholdVAD thu SpeechlyClient. Preliminarily added UseCouldSpeechProcessing flag.
