+++ speechly-client-net-standard-2.0/Speechly/SLUClient/Types.cs
@@ -102,7 +102,7 @@ namespace Speechly.SLUClient {
   public class Platform {
     public static string GetPersistentStoragePath()
     {
-#if UNITY_EDITOR_OSX || UNITY_STANDALONE_OSX || UNITY_EDITOR_WIN || UNITY_STANDALONE_WIN || UNITY_ANDROID || UNITY_IOS
+#if UNITY_EDITOR || UNITY_STANDALONE || UNITY_ANDROID || UNITY_IOS
       return UnityEngine.Application.persistentDataPath;
 #else
       return System.Environment.GetFolderPath(System.Environment.SpecialFolder.ApplicationData);
