+++ speechly-unity/Assets/Speechly/MicToSpeechly.cs
@@ -36,36 +36,9 @@ public class MicToSpeechly : MonoBehaviour
   public int HistoryFrames = 8;
   public bool CalcAudioPeaks = true;
   [Tooltip("Voice Activity Detection (VAD) using adaptive energy thresholding. Automatically controls listening based on audio 'loudness'.")]
-  public bool EnergyThresholdVAD = false;
-  [Range(0.0f, 1.0f)]
-  [Tooltip("Energy threshold - below this won't trigger activation")]
-  public float VADMinimumEnergy = 0.005f;
-  [Range(1.0f, 10.0f)]
-  [Tooltip("Signal-to-noise energy ratio needed for frame to be 'loud'")]
-  public float VADSignalToNoise = 2.0f;
-  [Range(1, 32)]
-  [Tooltip("Number of past frames analyzed for energy threshold VAD. Should be <= than HistoryFrames.")]
-  public int VADFrames = 5;
-  [Range(.0f, 1.0f)]
-  [Tooltip("Minimum 'loud' to 'silent' frame ratio in history to activate 'IsSignalDetected'")]
-  public float VADActivation = 0.7f;
-  [Range(.0f, 1.0f)]
-  [Tooltip("Maximum 'loud' to 'silent' frame ratio in history to inactivate 'IsSignalDetected'. Only evaluated when the sustain period is over.")]
-  public float VADRelease = 0.2f;
-  [Range(0, 8000)]
-  [Tooltip("Duration to keep 'IsSignalDetected' active. Renewed as long as VADActivation is holds true.")]
-  public int VADSustainMillis = 3000;
-  [Range(0, 5000)]
-  [Tooltip("Rate of background noise learn. Defined as duration in which background noise energy is moved halfway towards current frame's energy.")]
-  public int VADNoiseHalftimeMillis = 400;
-  [Tooltip("Disable VAD listening control if you want to use the energy threshold but want to implement custom listening control by reading IsSignalDetected state.")]
-  public bool VADControlListening = true;
+  public EnergyTresholdVAD Vad = new EnergyTresholdVAD{ Enabled = false };
   public bool DebugPrint = false;
   public float Peak {get; private set; } = 0f;
-  public float Energy {get; private set; } = 0f;
-  public float BaselineEnergy {get; private set; } = -1f;
-  public bool IsSignalDetected {get; private set; }
-  private int loudFrameBits = 0;
   public SpeechlyClient SpeechlyClient { get; private set; }
   private AudioClip clip;
   private float[] waveData;
@@ -76,7 +49,6 @@ public class MicToSpeechly : MonoBehaviour
 
   private int historySizeSamples;
   private int frameSamples;
-  private float vadSustainMillisLeft = 0;
   private bool wasVADEnabled = false;
 
   private void Awake() 
@@ -95,6 +67,7 @@ public class MicToSpeechly : MonoBehaviour
       apiUrl: SpeechlyEnv == SpeechlyEnvironment.Production ? null : "wss://staging.speechly.com/ws/v1?sampleRate=16000",
       appId: this.AppId,
       deviceId: SystemInfo.deviceUniqueIdentifier,
+      vad: this.Vad,
       manualUpdate: true,
       debug: DebugPrint
     );
@@ -175,24 +148,22 @@ public class MicToSpeechly : MonoBehaviour
           int audioPos = effectiveHistorySamples;
 
           while (unprocessedSamplesLeft >= frameSamples) {
-            // AnalyzeAudioFrame(in waveData, audioPos, frameSamples);
+            // If listening, send audio
+            if (SpeechlyClient.SamplesSent == 0) {
+              task = SpeechlyClient.ProcessFrame(waveData, 0, audioPos + frameSamples);
+            } else {
+              task = SpeechlyClient.ProcessFrame(waveData, audioPos, frameSamples);
+            }
+            yield return new WaitUntil(() => task.IsCompleted);
+            audioSent = true;
 
             // Control listening
-            if (EnergyThresholdVAD && VADControlListening) {
-              ControlListening(IsSignalDetected);
-            } else {
-              EnsureStopContext();
-            }
-
-            // If listening, send audio
-            if (SpeechlyClient.IsListening) {
-              if (SpeechlyClient.SamplesSent == 0) {
-                task = SpeechlyClient.ProcessFrame(waveData, 0, audioPos + frameSamples);
+            if (Vad != null) {
+              if (Vad.Enabled && Vad.VADControlListening) {
+                wasVADEnabled = true;
               } else {
-                task = SpeechlyClient.ProcessFrame(waveData, audioPos, frameSamples);
+                EnsureStopContext();
               }
-              audioSent = true;
-              yield return new WaitUntil(() => task.IsCompleted);
             }
 
             // Next frame
@@ -210,96 +181,13 @@ public class MicToSpeechly : MonoBehaviour
     }
   }
 
-  private void AnalyzeAudioFrame(in float[] waveData, int s, int frameSamples) {
-    if (CalcAudioPeaks) {
-      Peak = Mathf.Max(Peak, AudioTools.GetAudioPeak(in waveData, s, frameSamples));
-    }
-
-    if (EnergyThresholdVAD) {
-      Energy = AudioTools.GetEnergy(in waveData, s, frameSamples);
-
-      if (BaselineEnergy < 0f) {
-        BaselineEnergy = Energy;
-      }
-      bool isLoudFrame = Energy > Math.Max(VADMinimumEnergy, BaselineEnergy * VADSignalToNoise);
-      PushFrameHistory(isLoudFrame);
-      IsSignalDetected = DetermineNewSignalState(IsSignalDetected);
-      AdaptBackgroundNoise();
-    }
-  }
-
-  private void ControlListening(bool newState) {
-    wasVADEnabled = true;
-
-    if (!SpeechlyClient.IsListening && newState) {
-      StartContext();
-    }
-
-    if (SpeechlyClient.IsListening && !newState) {
-      StopContext();
-    }
-  }
-
   private void EnsureStopContext() {
     // Turn off listening when VAD is disabled
     if (wasVADEnabled) {
       wasVADEnabled = false;
-      ResetVAD();
       if (SpeechlyClient.IsListening) {
         StopContext();
       }
-    }
-  }
-
-  private bool DetermineNewSignalState(bool currentState) {
-    vadSustainMillisLeft = Math.Max(vadSustainMillisLeft - FrameMillis, 0);
-
-    int loudFrames = CountLoudFrames(VADFrames);
-    float loudFrameRatio = (1f * loudFrames) / VADFrames;
-
-    if (loudFrameRatio >= VADActivation) {
-      vadSustainMillisLeft = VADSustainMillis;
-      return true;
-    }
-
-    if (loudFrameRatio < VADRelease && vadSustainMillisLeft == 0) {
-      return false;
-    }
-
-    return currentState;
-  }
-
-  private void AdaptBackgroundNoise() {
-    // Gradually learn background noise level
-    if (!IsSignalDetected) {
-      if (VADNoiseHalftimeMillis > 0f) {
-        var decay = (float)Math.Pow(2.0, -FrameMillis / (double)VADNoiseHalftimeMillis);
-        BaselineEnergy = (BaselineEnergy * decay) + (Energy * (1f - decay));
-      }
-    }
-  }
-
-  private void PushFrameHistory(bool isLoud) {
-    loudFrameBits = (isLoud ? 1 : 0) | (loudFrameBits << 1);
-  }
-
-  private int CountLoudFrames(int numHistoryFrames) {
-    int numActiveFrames = 0;
-    int t = loudFrameBits;
-    while (numHistoryFrames > 0) {
-      if ((t & 1) == 1) numActiveFrames++;
-      t = t >> 1;
-      numHistoryFrames--;
-    }
-    return numActiveFrames;
-  }
-
-  private void ResetVAD() {
-    if (!EnergyThresholdVAD) {
-      IsSignalDetected = false;
-      loudFrameBits = 0;
-      Energy = 0f;
-      BaselineEnergy = -1f;
     }
   }
